refactor vacuum utilities
