+++ step-generation/src/utils/vacuumPythonArgs/__tests__/getVacuumProfileStepString.test.ts
@@ -1,35 +1,8 @@
 import { describe, expect, it } from 'vitest'
 
-import {
-  getVacuumProfileStepString,
-  getVacuumPumpHoldArgsPython,
-} from '../vacuumPythonArgs'
+import { getVacuumProfileStepString } from '../getVacuumProfileStepString'
 
 import type { VacuumProfile } from '@opentrons/shared-data'
-
-describe('getVacuumPumpHoldArgsPython', () => {
-  it('emits duration only when ventAfter is omitted', () => {
-    expect(getVacuumPumpHoldArgsPython(30)).toEqual(['duration=30'])
-  })
-
-  it('emits duration and vent_after when ventAfter is true', () => {
-    expect(getVacuumPumpHoldArgsPython(10, true)).toEqual([
-      'duration=10',
-      'vent_after=True',
-    ])
-  })
-
-  it('emits duration and vent_after when ventAfter is false', () => {
-    expect(getVacuumPumpHoldArgsPython(10, false)).toEqual([
-      'duration=10',
-      'vent_after=False',
-    ])
-  })
-
-  it('formats non-integer duration like Python numbers', () => {
-    expect(getVacuumPumpHoldArgsPython(1.5)).toEqual(['duration=1.5'])
-  })
-})
 
 describe('getVacuumProfileStepString', () => {
   const indentedPressureStep =
