+++ step-generation/src/utils/vacuumPythonArgs/getVacuumPumpHoldArgsPython.ts
@@ -1,0 +1,11 @@
+import { formatPyValue } from '../pythonFormat'
+
+export const getVacuumPumpHoldArgsPython = (
+  duration: number,
+  ventAfter?: boolean
+): string[] => {
+  return [
+    `duration=${formatPyValue(duration)}`,
+    ...(ventAfter != null ? [`vent_after=${formatPyValue(ventAfter)}`] : []),
+  ]
+}
